Changed file and model Structure
