+++ Cargo.toml
@@ -11,6 +11,7 @@ rfd = "0.8.0"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.79"
 tokio = { version = "1.17.0", features = ["full"] }
+lazy-static-include = "3.1.1"
 
 [profile.release]
 lto = true
