+++ src/lib/mod.rs
@@ -1,2 +1,2 @@
 pub mod gui;
-pub mod model;
+pub mod models;
