+++ src/lib/models/account.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Account {
+    pub username: String,
+    pub password: String,
+    pub premium: bool,
+}
